project2RAGLangraph/main.py
# 
from dotenv import load_dotenv

load_dotenv()

from graph.graph import app

if __name__ == "__main__":
    print("Hello Advanced Corrective RAG")
    print(app.invoke(input={"question": "what is agent memory?"}))