project2RAGLangraph/pyproject.toml
[project]
name = "project2raglangraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.1.0",
    "chromadb>=1.5.1",
    "isort>=8.0.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.9",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
